--- src/const.js
@@ -1,8 +1,0 @@
-const FilterType = {
-  EVERYTHING: 'everything',
-  FUTURE: 'future',
-  PRESENT: 'present',
-  PAST: 'past'
-};
-
-export {FilterType};
