+++ src/utils/filter.js
@@ -1,5 +1,11 @@
-import {FilterType} from '../const.js';
-import {isPointPast, isPointFuture, isPointPresent} from '../mock/point.js';
+import dayjs from 'dayjs';
+
+const FilterType = {
+  EVERYTHING: 'everything',
+  FUTURE: 'future',
+  PRESENT: 'present',
+  PAST: 'past'
+};
 
 const filter = {
   [FilterType.EVERYTHING]: (points) => [...points],
@@ -8,4 +14,27 @@ const filter = {
   [FilterType.PAST]: (points) => points.filter((point) => isPointPast(point)),
 };
 
-export {filter};
+function generateFilters(points) {
+  return Object.entries(filter)
+    .map(([filterType, filterPoints]) => ({
+      type: filterType,
+      hasPoints: filterPoints(points).length > 0
+    }));
+}
+
+
+function isPointFuture(point) {
+  return dayjs().isBefore(point.dateFrom);
+}
+
+function isPointPresent(point) {
+  return (dayjs().isAfter(point.dateFrom) && dayjs().isBefore(point.dateTo));
+}
+
+function isPointPast(point) {
+  return dayjs().isAfter(point.dateTo);
+}
+
+
+export {generateFilters, FilterType, filter, isPointFuture, isPointPresent, isPointPast};
+
