+++ src/view/sort-view.js
@@ -1,32 +1,29 @@
 import AbstractView from '../framework/view/abstract-view.js';
 
+const DEFAUL_FILTERS = [
+  {id: 'day', text: 'Day', checked: true},
+  {id: 'event', text: 'Event', disabled: true},
+  {id: 'time', text: 'Time'},
+  {id: 'price', text: 'Price'},
+  {id: 'offer', text: 'Offers', disabled: true}
+];
+
+function createSortItem({id, text, checked, disabled} = DEFAUL_FILTERS) {
+
+
+  return `<div class="trip-sort__item  trip-sort__item--${id}">
+  <input id="sort-${id}" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-${id}" ${checked ? 'checked' : ''} ${disabled ? 'disabled' : ''}>
+  <label class="trip-sort__btn" for="sort-${id}">${text}</label>
+</div>`;
+}
+
+
 function createSortViewTemplate () {
   return `
   <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-  <div class="trip-sort__item  trip-sort__item--day">
-    <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
-    <label class="trip-sort__btn" for="sort-day">Day</label>
-  </div>
 
-  <div class="trip-sort__item  trip-sort__item--event">
-    <input id="sort-event" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-event" disabled>
-    <label class="trip-sort__btn" for="sort-event">Event</label>
-  </div>
+    ${DEFAUL_FILTERS.map(createSortItem).join('')}
 
-  <div class="trip-sort__item  trip-sort__item--time">
-    <input id="sort-time" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-time">
-    <label class="trip-sort__btn" for="sort-time">Time</label>
-  </div>
-
-  <div class="trip-sort__item  trip-sort__item--price">
-    <input id="sort-price" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-price">
-    <label class="trip-sort__btn" for="sort-price">Price</label>
-  </div>
-
-  <div class="trip-sort__item  trip-sort__item--offer">
-    <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
-    <label class="trip-sort__btn" for="sort-offer">Offers</label>
-  </div>
 </form>
 `;
 }
